+++ dytnsapi-20200217/alibabacloud_dytnsapi20200217/models.py
@@ -4844,6 +4844,7 @@ class PhoneNumberEncryptRequest(TeaModel):
         auth_code: str = None,
         input_number: str = None,
         mask: str = None,
+        out_id: str = None,
         owner_id: int = None,
         resource_owner_account: str = None,
         resource_owner_id: int = None,
@@ -4866,6 +4867,7 @@ class PhoneNumberEncryptRequest(TeaModel):
         # 
         # This parameter is required.
         self.mask = mask
+        self.out_id = out_id
         self.owner_id = owner_id
         self.resource_owner_account = resource_owner_account
         self.resource_owner_id = resource_owner_id
@@ -4885,6 +4887,8 @@ class PhoneNumberEncryptRequest(TeaModel):
             result['InputNumber'] = self.input_number
         if self.mask is not None:
             result['Mask'] = self.mask
+        if self.out_id is not None:
+            result['OutId'] = self.out_id
         if self.owner_id is not None:
             result['OwnerId'] = self.owner_id
         if self.resource_owner_account is not None:
@@ -4901,6 +4905,8 @@ class PhoneNumberEncryptRequest(TeaModel):
             self.input_number = m.get('InputNumber')
         if m.get('Mask') is not None:
             self.mask = m.get('Mask')
+        if m.get('OutId') is not None:
+            self.out_id = m.get('OutId')
         if m.get('OwnerId') is not None:
             self.owner_id = m.get('OwnerId')
         if m.get('ResourceOwnerAccount') is not None:
@@ -4916,6 +4922,7 @@ class PhoneNumberEncryptResponseBodyData(TeaModel):
         encrypted_number: str = None,
         expire_time: str = None,
         original_number: str = None,
+        out_id: str = None,
     ):
         # The encrypted phone number.
         self.encrypted_number = encrypted_number
@@ -4923,6 +4930,7 @@ class PhoneNumberEncryptResponseBodyData(TeaModel):
         self.expire_time = expire_time
         # The original phone number.
         self.original_number = original_number
+        self.out_id = out_id
 
     def validate(self):
         pass
@@ -4939,6 +4947,8 @@ class PhoneNumberEncryptResponseBodyData(TeaModel):
             result['ExpireTime'] = self.expire_time
         if self.original_number is not None:
             result['OriginalNumber'] = self.original_number
+        if self.out_id is not None:
+            result['OutId'] = self.out_id
         return result
 
     def from_map(self, m: dict = None):
@@ -4949,6 +4959,8 @@ class PhoneNumberEncryptResponseBodyData(TeaModel):
             self.expire_time = m.get('ExpireTime')
         if m.get('OriginalNumber') is not None:
             self.original_number = m.get('OriginalNumber')
+        if m.get('OutId') is not None:
+            self.out_id = m.get('OutId')
         return self
 
 
@@ -5950,211 +5962,6 @@ class PhoneNumberStatusForSmsResponse(TeaModel):
             self.status_code = m.get('statusCode')
         if m.get('body') is not None:
             temp_model = PhoneNumberStatusForSmsResponseBody()
-            self.body = temp_model.from_map(m['body'])
-        return self
-
-
-class PhoneNumberStatusForVirtualRequest(TeaModel):
-    def __init__(
-        self,
-        auth_code: str = None,
-        input_number: str = None,
-        mask: str = None,
-        owner_id: int = None,
-        resource_owner_account: str = None,
-        resource_owner_id: int = None,
-    ):
-        # The authorization code.
-        # 
-        # >  On the **My Applications** page in the [Cell Phone Number Service console](https://dytns.console.aliyun.com/analysis/apply), you can obtain the authorization ID.
-        # 
-        # This parameter is required.
-        self.auth_code = auth_code
-        # The phone number to be queried.
-        # 
-        # *   If the value of Mask is NORMAL, the value of this field is an 11-digit phone number.
-        # *   If the value of Mask is MD5, the value of this field is a 32-bit encrypted string.
-        # *   If the value of Mask is SHA256, the value of this field is a 64-bit encrypted string.
-        # 
-        # >  Letters in the encrypted strings are not case-sensitive.
-        # 
-        # This parameter is required.
-        self.input_number = input_number
-        # The encryption method of the phone number. Valid values:
-        # 
-        # *   **NORMAL**: The phone number is not encrypted.
-        # *   **MD5**\
-        # *   **SHA256**\
-        # 
-        # This parameter is required.
-        self.mask = mask
-        self.owner_id = owner_id
-        self.resource_owner_account = resource_owner_account
-        self.resource_owner_id = resource_owner_id
-
-    def validate(self):
-        pass
-
-    def to_map(self):
-        _map = super().to_map()
-        if _map is not None:
-            return _map
-
-        result = dict()
-        if self.auth_code is not None:
-            result['AuthCode'] = self.auth_code
-        if self.input_number is not None:
-            result['InputNumber'] = self.input_number
-        if self.mask is not None:
-            result['Mask'] = self.mask
-        if self.owner_id is not None:
-            result['OwnerId'] = self.owner_id
-        if self.resource_owner_account is not None:
-            result['ResourceOwnerAccount'] = self.resource_owner_account
-        if self.resource_owner_id is not None:
-            result['ResourceOwnerId'] = self.resource_owner_id
-        return result
-
-    def from_map(self, m: dict = None):
-        m = m or dict()
-        if m.get('AuthCode') is not None:
-            self.auth_code = m.get('AuthCode')
-        if m.get('InputNumber') is not None:
-            self.input_number = m.get('InputNumber')
-        if m.get('Mask') is not None:
-            self.mask = m.get('Mask')
-        if m.get('OwnerId') is not None:
-            self.owner_id = m.get('OwnerId')
-        if m.get('ResourceOwnerAccount') is not None:
-            self.resource_owner_account = m.get('ResourceOwnerAccount')
-        if m.get('ResourceOwnerId') is not None:
-            self.resource_owner_id = m.get('ResourceOwnerId')
-        return self
-
-
-class PhoneNumberStatusForVirtualResponseBodyData(TeaModel):
-    def __init__(
-        self,
-        is_privacy_number: bool = None,
-    ):
-        # Indicate whether the phone number is a virtual number assigned by the carrier. Valid values:
-        # 
-        # *   **true**\
-        # *   **false**\
-        self.is_privacy_number = is_privacy_number
-
-    def validate(self):
-        pass
-
-    def to_map(self):
-        _map = super().to_map()
-        if _map is not None:
-            return _map
-
-        result = dict()
-        if self.is_privacy_number is not None:
-            result['IsPrivacyNumber'] = self.is_privacy_number
-        return result
-
-    def from_map(self, m: dict = None):
-        m = m or dict()
-        if m.get('IsPrivacyNumber') is not None:
-            self.is_privacy_number = m.get('IsPrivacyNumber')
-        return self
-
-
-class PhoneNumberStatusForVirtualResponseBody(TeaModel):
-    def __init__(
-        self,
-        code: str = None,
-        data: PhoneNumberStatusForVirtualResponseBodyData = None,
-        message: str = None,
-        request_id: str = None,
-    ):
-        # The response code. Valid values:
-        # 
-        # *   **OK**: The request is successful.
-        # *   **OperatorLimit**: The carrier prohibits the query of the phone number.
-        # *   **RequestFrequencyLimit**: Repeated queries for the same phone number at a high frequency within a short period of time are prohibited due to restrictions that are set by carriers. If this error code is returned, please try again later.
-        self.code = code
-        # The response parameters.
-        self.data = data
-        # The returned message.
-        self.message = message
-        # The request ID.
-        self.request_id = request_id
-
-    def validate(self):
-        if self.data:
-            self.data.validate()
-
-    def to_map(self):
-        _map = super().to_map()
-        if _map is not None:
-            return _map
-
-        result = dict()
-        if self.code is not None:
-            result['Code'] = self.code
-        if self.data is not None:
-            result['Data'] = self.data.to_map()
-        if self.message is not None:
-            result['Message'] = self.message
-        if self.request_id is not None:
-            result['RequestId'] = self.request_id
-        return result
-
-    def from_map(self, m: dict = None):
-        m = m or dict()
-        if m.get('Code') is not None:
-            self.code = m.get('Code')
-        if m.get('Data') is not None:
-            temp_model = PhoneNumberStatusForVirtualResponseBodyData()
-            self.data = temp_model.from_map(m['Data'])
-        if m.get('Message') is not None:
-            self.message = m.get('Message')
-        if m.get('RequestId') is not None:
-            self.request_id = m.get('RequestId')
-        return self
-
-
-class PhoneNumberStatusForVirtualResponse(TeaModel):
-    def __init__(
-        self,
-        headers: Dict[str, str] = None,
-        status_code: int = None,
-        body: PhoneNumberStatusForVirtualResponseBody = None,
-    ):
-        self.headers = headers
-        self.status_code = status_code
-        self.body = body
-
-    def validate(self):
-        if self.body:
-            self.body.validate()
-
-    def to_map(self):
-        _map = super().to_map()
-        if _map is not None:
-            return _map
-
-        result = dict()
-        if self.headers is not None:
-            result['headers'] = self.headers
-        if self.status_code is not None:
-            result['statusCode'] = self.status_code
-        if self.body is not None:
-            result['body'] = self.body.to_map()
-        return result
-
-    def from_map(self, m: dict = None):
-        m = m or dict()
-        if m.get('headers') is not None:
-            self.headers = m.get('headers')
-        if m.get('statusCode') is not None:
-            self.status_code = m.get('statusCode')
-        if m.get('body') is not None:
-            temp_model = PhoneNumberStatusForVirtualResponseBody()
             self.body = temp_model.from_map(m['body'])
         return self